Changing doors' notations
Changing doors' notation in order to only have NOT, OR, NOR, AND, NAND, XOR and XNOR gates. Removing the possibility to change doors' notation. I do that to prepare the printing of doors' notation with logic gates symbols.  I also tried to make the trees' depth as small as possible so that the logic gates' symbols can fit inside or the door's window.